requirements.txt
fastapi
uvicorn[standard]
pydantic
chromadb
openai
python-multipart
tqdm
pdfplumber
python-dotenv
streamlit
requests
